cleaned up unused dependencies and taskt.Updater
