+++ taskt-updater/Folders.cs
@@ -1,0 +1,14 @@
+﻿namespace taskt.Updater
+{
+    public class Folders
+    {
+        public string Source { get; private set; }
+        public string Target { get; private set; }
+
+        public Folders(string source, string target)
+        {
+            Source = source;
+            Target = target;
+        }
+    }
+}
