+++ taskt.Commands/packages.config.new.20200827225752
@@ -1,39 +1,17 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<!--
-Copyright (c) 2017 Jason Bayldon
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
 <packages>
-  <package id="HtmlAgilityPack" version="1.11.24" targetFramework="net48" />
-  <package id="ImapX" version="2.0.0.18" targetFramework="net452" />
-  <package id="InputSimulator" version="1.0.4.0" targetFramework="net48" />
-  <package id="JetBrains.Annotations" version="2020.1.0" targetFramework="net48" />
   <package id="MailKit" version="2.8.0" targetFramework="net48" />
   <package id="Microsoft.AspNet.SignalR.Client" version="2.4.1" targetFramework="net48" />
   <package id="Microsoft.AspNet.SignalR.Core" version="2.4.1" targetFramework="net48" />
-  <package id="Microsoft.Office.Interop.Excel" version="15.0.4795.1000" targetFramework="net48" />
   <package id="Microsoft.Office.Interop.Outlook" version="15.0.4797.1003" targetFramework="net48" />
   <package id="Microsoft.Office.Interop.Word" version="15.0.4797.1003" targetFramework="net48" />
   <package id="Microsoft.Owin" version="3.0.1" targetFramework="net48" />
   <package id="Microsoft.Owin.Security" version="3.0.1" targetFramework="net48" />
   <package id="MimeKit" version="2.9.1" targetFramework="net48" />
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net452" />
-  <package id="OneNoteOCR" version="1.0.0.0" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net48" />
   <package id="Owin" version="1.0" targetFramework="net48" />
   <package id="Portable.BouncyCastle" version="1.8.5" targetFramework="net48" />
-  <package id="RestSharp" version="106.6.9" targetFramework="net452" />
-  <package id="Selenium.Support" version="3.141.0" targetFramework="net48" />
+  <package id="RestSharp" version="106.6.9" targetFramework="net48" />
   <package id="Selenium.WebDriver" version="3.141.0" targetFramework="net48" />
   <package id="Serilog" version="2.9.0" targetFramework="net48" />
   <package id="Serilog.Formatting.Compact" version="1.1.0" targetFramework="net48" />
@@ -43,7 +21,4 @@ limitations under the License.
   <package id="Serilog.Sinks.RollingFile" version="3.3.0" targetFramework="net48" />
   <package id="Serilog.Sinks.SignalR" version="2.0.10" targetFramework="net48" />
   <package id="SharpSimpleNLG" version="1.2.1" targetFramework="net48" />
-  <package id="SuperSocket.ClientEngine.Core" version="0.8.0.13" targetFramework="net452" />
-  <package id="TaskScheduler" version="2.6.1" targetFramework="net452" />
-  <package id="WebSocket4Net" version="0.15.0" targetFramework="net452" />
 </packages>
