+++ taskt.Studio/UI/Forms/UIForm.cs
@@ -45,7 +45,7 @@ namespace taskt.UI.Forms
 
             // UIForm
             ClientSize = new Size(284, 261);
-            Icon = ((Icon)(resources.GetObject("$Icon")));
+            Icon = (Icon)resources.GetObject("$this.Icon");
             Name = "UIForm";
             Load += new EventHandler(UIForm_Load);
             ResumeLayout(false);
