+++ taskt.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 16.0.29613.14
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "taskt.Studio", "taskt.Studio\taskt.Studio.csproj", "{C1BE3204-94D1-4A9A-AE30-C3E302383182}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "taskt-updater", "taskt-updater\taskt-updater.csproj", "{3FC72D0C-D083-44C4-96E9-13FE599F8E0E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "taskt.Updater", "taskt-updater\taskt.Updater.csproj", "{3FC72D0C-D083-44C4-96E9-13FE599F8E0E}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "taskt.Core", "taskt.Core\taskt.Core.csproj", "{6956836D-D2B0-4B7A-B06D-FA3A22B4C4FF}"
 EndProject
